% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{clusterTimedTree}
\alias{clusterTimedTree}
\title{find clusters in timed tree}
\usage{
clusterTimedTree(timedTree, K_cutoff, beta, perc_cutoff, maxK = 25)
}
\arguments{
\item{timedTree}{timed tree in newick format}

\item{beta}{transmission rate}

\item{perc_cutoff}{cut-off percentage}

\item{maxK}{max number of transmissions considered}

\item{k_cutoff}{number of transmissions in cutoff criterion}
}
\description{
find clusters in timed tree
}
\examples{
findTreeCuts(branchtree, 10, 2.2, 0.2)
}
