% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{compareClusters}
\alias{compareClusters}
\title{compare similarity of clusters}
\usage{
compareClusters(SNPClusters, transClusters, roundTo = 3)
}
\arguments{
\item{SNPClusters}{List of clusters}

\item{transClusters}{List of clusters}

\item{roundTo}{Number of decimal places in return matrix}
}
\description{
compare similarity of clusters
}
