% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{createModel}
\alias{createModel}
\title{create model to be used for cluster analysis}
\usage{
createModel(ids = NULL, dates = NULL, snpMatrix = matrix(, 0, 0),
  lambda = 1, beta = 1, percCutOff = 0.2, thSNP = seq(1, 10),
  thTrans = seq(1, 10))
}
\arguments{
\item{ids}{(optional) vector of unique sample IDs}

\item{dates}{(optional) vector of dates (formatted as numbers)}

\item{snpMatrix}{(optional) symmetric matrix of SNP distances supplied in the same order as ids}

\item{lambda}{(optional) clock rate}

\item{beta}{(optional) transmission rate}

\item{percCutOff}{(optional) cut-off percentage}

\item{thSNP}{(optional) SNP threshold values as a vector of integers}

\item{thTrans}{(optional) transmission threshold values as a vector of integers}
}
\value{
model object for further analysis
}
\description{
create model to be used for cluster analysis
}
\examples{
myModel <- createModel()
}
