% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{cutoffLevel}
\alias{cutoffLevel}
\title{create levels at which number of transmissions passes the threshold}
\usage{
cutoffLevel(t, maxK, beta, perc_cutoff)
}
\arguments{
\item{maxK}{number of transmissions}

\item{beta}{transmission rate}

\item{perc_cutoff}{cut-off percentage}
}
\value{
time length of cutoff
}
\description{
create levels at which number of transmissions passes the threshold
}
