% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{makeSNPClusters}
\alias{makeSNPClusters}
\title{make SNP-based clusters for each threshold set in the model}
\usage{
makeSNPClusters(thisModel, nameRoot = "SNPcluster", writeFile = TRUE)
}
\arguments{
\item{thisModel}{cluster analysis model created using createModel()}

\item{nameRoot}{(optional) basis of name to be used for output cluster files}

\item{writeFile}{(optional) cluster files will be written if TRUE}
}
\value{
list of clusters
}
\description{
make SNP-based clusters for each threshold set in the model
}
\examples{
mySNPClusters <- makeSNPClusters(myModel, 'test')
}
