% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{makeTransClusters}
\alias{makeTransClusters}
\title{make transmission-based clusters for each threshold set in the model}
\usage{
makeTransClusters(thisModel, nameBase = "transcluster",
  writeFile = TRUE)
}
\arguments{
\item{thisModel}{cluster analysis model created using createModel()}

\item{nameBase}{(optional) basis of name to be used for output cluster files}

\item{writeFile}{(optional) cluster files will be written if TRUE}
}
\value{
list of clusters
}
\description{
make transmission-based clusters for each threshold set in the model
}
\examples{
myTransClusters <- makeTransClusters(myModel, 'test')
}
\keyword{cluster}
