% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resist.R
\name{makeTransClustersResist}
\alias{makeTransClustersResist}
\title{make transmission-based clusters for each threshold set in the model using resistance SNPs}
\usage{
makeTransClustersResist(thisModel, nameBase = "transcluster",
  writeFile = TRUE)
}
\arguments{
\item{thisModel}{cluster analysis model created using createModel()}

\item{nameBase}{basis of name to be used for output cluster files}

\item{writeFile}{cluster files will be written if TRUE}
}
\value{
list of clusters
}
\description{
make transmission-based clusters for each threshold set in the model using resistance SNPs
}
\examples{
myTransClusters <- makeTransClustersResist(myModel, 'test')
}
\keyword{cluster}
