% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{nTransCutoff}
\alias{nTransCutoff}
\title{create levels at which number of transmissions passes the threshold}
\usage{
nTransCutoff(N, t1, t2, lambda, beta, perc_cutoff, maxN = 25, maxK = 25)
}
\arguments{
\item{N}{SNP distance}

\item{t1}{date of first sample}

\item{t2}{date of second sample}

\item{lambda}{clock rate}

\item{beta}{transmission rate}

\item{perc_cutoff}{cut-off percentage}

\item{maxN}{for optimisation, SNP distance greater than this considered not in same cluster}
}
\value{
number of transmissions reached
}
\description{
create levels at which number of transmissions passes the threshold
}
