% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotSNPClusters}
\alias{plotSNPClusters}
\title{interactive plot of SNP clusters with weighted edges}
\usage{
plotSNPClusters(clus, myModel, eWidth = 2, vSize = 5, vFontSize = 0.2,
  vColor = "cyan", level = 9, thick = 1)
}
\arguments{
\item{clus}{clusters obtained from makeSNPClusters()}

\item{myModel}{cluster analyis model created using createModel()}

\item{eWidth}{width of edges}

\item{vSize}{size of vertices}

\item{vFontSize}{font size of vertex labels}

\item{vColor}{vertex colour}

\item{level}{SNP cut-off level}

\item{thick}{edge weight adjustment factor}
}
\value{
graph object
}
\description{
interactive plot of SNP clusters with weighted edges
}
\examples{
plotSNPClusters(fake, bcModel, level=4, vColor='orange', vSize=4, thick=1.25)
}
