% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotTransClusters}
\alias{plotTransClusters}
\title{plot of transmission clusters with weighted edges}
\usage{
plotTransClusters(myModel, eWidth = 2, vSize = 5, vFontSize = 1,
  vColor = "cyan", level = 1, thick = 1, labelOffset = 1,
  showLabels = TRUE)
}
\arguments{
\item{myModel}{cluster analyis model created using createModel()}

\item{eWidth}{width of edges}

\item{vSize}{size of vertices}

\item{vFontSize}{font size of vertex labels}

\item{vColor}{vertex colour}

\item{level}{transmission cut-off level}

\item{thick}{edge weight adjustment factor}

\item{labelOffset}{offset height for vertex labels}
}
\value{
graph object
}
\description{
plot of transmission clusters with weighted edges
}
\examples{
bcg <- plotTransClusters(clus, bcModel, level=3, vColor='lightblue',vSize=4, thick = 2, vFontSize=1)
}
