% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{regionColour}
\alias{regionColour}
\title{return colour label}
\usage{
regionColour(thisModel, regionLabel)
}
\arguments{
\item{regionLabel}{region label}
}
\value{
colour label
}
\description{
return colour label
}
