% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{regionWeight}
\alias{regionWeight}
\title{return region weights}
\usage{
regionWeight(thisModel, i, j)
}
\arguments{
\item{thisModel}{Model object}

\item{i}{index}

\item{j}{index}
}
\value{
weight
}
\description{
return region weights
}
