% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resist.R
\name{setCutoffsResist}
\alias{setCutoffsResist}
\title{set levels at which number of transmissions passes the threshold}
\usage{
setCutoffsResist(thisModel, lambda_factor = 5, maxK = 25)
}
\arguments{
\item{thisModel}{the model}

\item{lambda_factor}{factor by which lambda is increased for resistance SNPs}

\item{maxK}{max number of transmissions looked for}
}
\value{
the model
}
\description{
set levels at which number of transmissions passes the threshold
}
