% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{setDatesFromFile}
\alias{setDatesFromFile}
\title{set sample dates from a file}
\usage{
setDatesFromFile(thisModel, filename)
}
\arguments{
\item{thisModel}{Model object}

\item{filename}{Name of the input date file}
}
\value{
thisModel is returned with the dates added to it
}
\description{
set sample dates from a file
}
\examples{
myModel <- setDatesFromFile(myModel, datefile)
}
