% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{setRegionColours}
\alias{setRegionColours}
\title{set region colours}
\usage{
setRegionColours(thisModel, colours = c("orange", "lightblue", "red",
  "blue", "purple", "lightgreen", "yellow", "green", "brown"))
}
\arguments{
\item{thisModel}{Model object}

\item{colours}{Vector of region colours}
}
\description{
set region colours
}
