% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{setRegionLabels}
\alias{setRegionLabels}
\title{set region labels}
\usage{
setRegionLabels(thisModel, labels = c("Region 1", "Region 2", "Region 3",
  "Region 4", "Region 5", "Region 6"))
}
\arguments{
\item{thisModel}{Model object}

\item{labels}{Vector of region labels}
}
\description{
set region labels
}
