% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{setRegionsFromFile}
\alias{setRegionsFromFile}
\title{set regions from a file}
\usage{
setRegionsFromFile(thisModel, filename)
}
\arguments{
\item{thisModel}{Model object}

\item{filename}{Name of the input date file}
}
\value{
thisModel is returned with the dates added to it
}
\description{
set regions from a file
}
