% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{setRegionWeights}
\alias{setRegionWeights}
\title{set region weightings}
\usage{
setRegionWeights(thisModel, sameRegion = 1, diffRegion = 0.1)
}
\arguments{
\item{thisModel}{Model object}

\item{sameRegion}{Weighting for cases from the same region}

\item{diffRegion}{Weighting for cases from the different regions}
}
\description{
set region weightings
}
