% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{setSNPFromFile}
\alias{setSNPFromFile}
\title{set SNP distance matrix from a file}
\usage{
setSNPFromFile(thisModel, filename)
}
\arguments{
\item{thisModel}{Model object}

\item{filename}{Name of the input SNP data file}
}
\description{
set SNP distance matrix from a file
}
\examples{
myModel <- setSNPFromFile(myModel, snpfile)
}
\keyword{SNP}
