% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resist.R
\name{setSNPResistFromFile}
\alias{setSNPResistFromFile}
\title{set SNP resistant distance matrix from a file}
\usage{
setSNPResistFromFile(thisModel, filename)
}
\arguments{
\item{thisModel}{Model object}

\item{filename}{Name of the input SNP data file}
}
\value{
the model
}
\description{
set SNP resistant distance matrix from a file
}
\examples{
myModel <- setSNPResistFromFile(myModel, snpfile)
}
\keyword{SNP}
