% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setSNPThresholds}
\alias{setSNPThresholds}
\title{set SNP thresholds for existing cluster analysis model}
\usage{
setSNPThresholds(thisModel, thSNP = seq(1, 10))
}
\arguments{
\item{thisModel}{cluster analsyis model created using createModel()}

\item{thSNP}{(optional) SNP threshold values as a vector of integers}
}
\value{
model object
}
\description{
set SNP thresholds for existing cluster analysis model
}
\examples{
myModel <- setSNPThresholds(myModel, seq(12))
}
