% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{setTransMatrixFromFile}
\alias{setTransMatrixFromFile}
\title{set transmission distance matrix from a file}
\usage{
setTransMatrixFromFile(thisModel, filename)
}
\arguments{
\item{thisModel}{Model object}

\item{filename}{Name of the input transmission data file}
}
\description{
set transmission distance matrix from a file
}
\examples{
myModel <- setTransMatrixFromFile(myModel, transfile)
}
\keyword{transmission}
