% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setTransThresholds}
\alias{setTransThresholds}
\title{set transmission thresholds for existing cluster analysis model}
\usage{
setTransThresholds(thisModel, thTrans = seq(1, 10))
}
\arguments{
\item{thisModel}{cluster analysis model created using createModel()}

\item{thTrans}{(optional) transmission threshold values as a vector of integers}
}
\value{
model object
}
\description{
set transmission thresholds for existing cluster analysis model
}
\examples{
myModel <- setTransThresholds(myModel, seq(12))
}
