% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{withinCluster}
\alias{withinCluster}
\title{test whether two cases are within the same transmission cluster}
\usage{
withinCluster(
  N,
  k_cutoff,
  t1,
  t2,
  lambda,
  beta,
  perc_cutoff,
  maxN = 25,
  N_resist = 0,
  lambda_resist = 0
)
}
\arguments{
\item{N}{SNP distance}

\item{k_cutoff}{number of transmissions below this threshold considered to be in same cluster}

\item{t1}{date of first sample}

\item{t2}{date of second sample}

\item{lambda}{clock rate}

\item{beta}{transmission rate}

\item{perc_cutoff}{cut-off percentage}

\item{maxN}{for optimisation, SNP distance greater than this considered not in same cluster}

\item{N_resist}{number of resistance-conferring SNPs as a subset of N}

\item{lambda_resist}{clock rate for resistance-conferring SNPs}
}
\value{
TRUE if within cluster, FALSE otherwise
}
\description{
test whether two cases are within the same transmission cluster
}
\examples{
isWithinCluster <- WithinCluster(N=5, k_cutoff=3, t1=2012.13, t2=2015.76, lambda=0.7, beta=1.3, perc_cutoff=0.2)
}
