% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotSNPClusters}
\alias{plotSNPClusters}
\title{plot of SNP clusters with weighted edges}
\usage{
plotSNPClusters(
  myModel,
  eWidth = 2,
  vSize = 5,
  vFontSize = 1,
  vColor = "cyan",
  level = 1,
  thick = 1,
  labelOffset = 1
)
}
\arguments{
\item{myModel}{cluster analyis model created using createModel()}

\item{eWidth}{width of edges}

\item{vSize}{size of vertices}

\item{vFontSize}{font size of vertex labels}

\item{vColor}{vertex colour}

\item{level}{SNP cut-off level}

\item{thick}{edge weight adjustment factor}

\item{labelOffset}{offset height for vertex labels}
}
\value{
graph object
}
\description{
plot of SNP clusters with weighted edges
}
\examples{
plotSNPClusters(clus, bcModel, level=4, vColor='orange', vSize=4, thick=1.25)
}
