% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visuals.R
\name{plotTransClusters}
\alias{plotTransClusters}
\title{interactive plot of transmission clusters with weighted edges}
\usage{
plotTransClusters(clus, myModel, eWidth = 2, vSize = 5, vFontSize = 1,
  vColor = "cyan", level = 10, thick = 1)
}
\arguments{
\item{clus}{clusters obtained from makeTransClusters()}

\item{myModel}{cluster analyis model created using createModel()}

\item{eWidth}{width of edges}

\item{vSize}{size of vertices}

\item{vFontSize}{font size of vertex labels}

\item{vColor}{vertex colour}

\item{level}{transmission SNP cut-off level}

\item{thick}{edge weight adjustment factor}
}
\value{
graph object
}
\description{
interactive plot of transmission clusters with weighted edges
}
\examples{
bcg <- plotTransClusters(fake, bcModel, level=3, vColor='lightblue',vSize=4, thick = 2, vFontSize=1)
}
