% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{setCutoffs}
\alias{setCutoffs}
\title{set levels at which number of transmissions passes the threshold}
\usage{
setCutoffs(thisModel, maxK = 25)
}
\arguments{
\item{thisModel}{the model}

\item{maxK}{max number of transmissions looked for}
}
\value{
the model
}
\description{
set levels at which number of transmissions passes the threshold
}
